test(docx): move the blank-spacer test to its own module
tests/test_backend_msword.py was 1496 lines and the new test pushed it
over the repository's 1500-line max-lines hook. Move the test into
tests/test_backend_msword_spacer.py and restore test_backend_msword.py
to its state on main. The local get_converter() helper is inlined as a
DocumentConverter built in the test; no test logic changes.
